refactor: inline temporal-shift helpers into DTensor value worker
