+++ nemo_rl/models/value/workers/dtensor_value_worker_v2.py
@@ -56,12 +56,40 @@ from nemo_rl.models.policy.workers.base_policy_worker import AbstractPolicyWorke
 from nemo_rl.models.policy.workers.patches import apply_transformer_engine_patch
 from nemo_rl.models.value.config import ValueConfig
 from nemo_rl.models.value.interfaces import ValueOutputSpec
-from nemo_rl.models.value.value_temporal_shift import (
-    RightShiftLossWrapper,
-    right_shift_values,
-)
 from nemo_rl.utils.checkpoint import CheckpointingConfig
 from nemo_rl.utils.nsys import wrap_with_nvtx_name
+
+
+def right_shift_values(values: torch.Tensor) -> torch.Tensor:
+    """Shift values right by 1 along the sequence dim (V(s_{t+1}) -> V(s_t)).
+
+    Aligns value predictions with the Megatron value worker convention so GAE
+    (rewards, returns), value targets, and value clipping all see the same
+    V(s_t) semantics across backends. Preserves the input tensor shape: the
+    first column becomes zeros and column t (t>=1) takes the value from
+    column t-1.
+    """
+    return torch.cat([torch.zeros_like(values[:, :1]), values[:, :-1]], dim=1)
+
+
+class RightShiftLossWrapper:
+    """Wrap a LossFunction so value logits are right-shifted before loss."""
+
+    def __init__(self, inner: LossFunction):
+        self._inner = inner
+
+    def __call__(self, *args, logits=None, **kwargs):
+        if logits is not None:
+            logits = right_shift_values(logits)
+            return self._inner(*args, logits=logits, **kwargs)
+        return self._inner(*args, **kwargs)
+
+    def __getattr__(self, name):
+        # __getattr__ is only called when normal attribute lookup fails, so
+        # _inner itself is still found through __dict__. Everything else
+        # (input_type, aggregation_type, etc.) delegates to the wrapped
+        # loss function.
+        return getattr(self._inner, name)
 
 
 @contextlib.contextmanager
